+++ data/impl/RecordSet.ts
@@ -11,7 +11,6 @@ import {maxBy, isNil} from 'lodash';
 import {StoreRecord, StoreRecordId} from '../StoreRecord';
 import {Store} from '../Store';
 import {Filter} from '../filter/Filter';
-import type {RsTransaction} from '@xh/hoist/data/Types';
 
 /**
  * Internal container for StoreRecord management within a Store.
@@ -177,7 +176,11 @@ export class RecordSet {
         return new RecordSet(this.store, recordMap);
     }
 
-    withTransaction(t: RsTransaction): RecordSet {
+    withTransaction(t: {
+        update?: StoreRecord[];
+        add?: StoreRecord[];
+        remove?: StoreRecordId[];
+    }): RecordSet {
         const {update, add, remove} = t;
 
         // Be sure to finalize any new records that are accepted.